refactor: fix circular deps (#500)
