+++ app/trpc/client.ts
@@ -1,12 +1,8 @@
 import { createFnRecordQueryProxy } from "@hiogawa/query-proxy";
 import { httpClientAdapter, proxyTinyRpc } from "@hiogawa/tiny-rpc";
 import { JSON_EXTRA } from "../utils/json-extra";
+import { RPC_ENDPOINT, RPC_GET_PATHS } from "./common";
 import { type rpcRoutes } from "./server";
-
-export const RPC_ENDPOINT = "/rpc";
-export const RPC_GET_PATHS: (keyof typeof rpcRoutes)[] = [
-  "videos_getCaptionEntries",
-];
 
 export const rpcClient = proxyTinyRpc<typeof rpcRoutes>({
   adapter: httpClientAdapter({
