+++ app/trpc/common.ts
@@ -1,0 +1,2 @@
+export const RPC_ENDPOINT = "/rpc";
+export const RPC_GET_PATHS = ["videos_getCaptionEntries" as const];
