+++ package.json
@@ -81,7 +81,7 @@
     "zod": "^3.22.2"
   },
   "devDependencies": {
-    "@hiogawa/icheck-ts": "0.0.1-pre.4",
+    "@hiogawa/icheck-ts": "0.0.1-pre.10",
     "@hiogawa/isort-ts": "1.1.2-pre.0",
     "@hiogawa/tiny-cli": "0.0.4-pre.1",
     "@hiogawa/unocss-preset-antd": "2.2.1-pre.3",
